HydroTek-22-PestDetection/pest-detection/pest-detection/requirements.txt
matplotlib>=3.2.2
numpy>=1.18.5
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
torchaudio
opencv-python>=4.1.2
tqdm>=4.41.0
flask
pandas
seaborn
Flask-RESTful
subprocess.run
gsutil
pymongo[srv]